Create a separate Scheduler abstraction.
